// src/app/core/enums/question-type.enum.ts
export enum QuestionType {
  MultipleChoice = 'MultipleChoice',
  MultipleResponse = 'MultipleResponse',
  TrueFalse = 'TrueFalse',
  FillInTheBlank = 'FillInTheBlank',
  Matching = 'Matching',
  Ordering = 'Ordering',
  Numeric = 'Numeric',
  ShortAnswer = 'ShortAnswer',
  Essay = 'Essay',
  FileUpload = 'FileUpload',
}